backend/requirements.txt
Django>=5.2.3
djangorestframework>=3.16.0
djangorestframework-simplejwt>=5.3.1
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.1
